@@ -18,18 +18,18 @@
     "@fortawesome/free-solid-svg-icons": "^7.0.1",
     "@fortawesome/react-fontawesome": "^3.0.2",
     "@tailwindcss/postcss": "^4.1.13",
-    "@types/node": "24.4.0",
+    "@types/node": "24.6.0",
     "@types/react-dom": "19.1.9",
-    "@types/react": "19.1.13",
+    "@types/react": "19.1.16",
     "@vercel/analytics": "^1.5.0",
-    "eslint-config-next": "15.5.3",
-    "eslint": "9.35.0",
-    "next": "15.5.3",
+    "eslint-config-next": "15.5.4",
+    "eslint": "9.36.0",
+    "next": "15.5.4",
     "postcss": "^8.5.6",
     "prettier": "^3",
     "react-dom": "19.1.1",
     "react": "19.1.1",
-    "sharp": "^0.34.3",
+    "sharp": "^0.34.4",
     "tailwindcss": "^4.1.13",
     "typescript": "5.9.2"
   }
